change min_amount_out and amount_received to integer types in sell shares flow, add Math.floor() to frontend sell value calculations, and update API error handling for integer validation

+++ backend/app/models/schemas.py
@@ -95,13 +95,13 @@ class SellSharesRequest(BaseModel):
     line_id: UUID
     outcome: Literal["yes", "no"]
     shares: float = Field(..., gt=0)
-    min_amount_out: float = Field(..., gt=0, description="Minimum GOOS to receive (slippage protection)")
+    min_amount_out: int = Field(..., gt=0, description="Minimum GOOS to receive (slippage protection)")
 
 
 class SellSharesResponse(BaseModel):
     """Response after selling shares."""
     shares_sold: float
-    amount_received: float  # GOOS received
+    amount_received: int  # GOOS received
     sell_price: float  # amount_received / shares_sold
     new_balance: int
     remaining_shares: float
